pyproject.toml
[project]
name = "support-copilot"
version = "0.1.0"
description = ""
authors = [
    {name = "raulwaard",email = "raulwarddev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "langgraph (>=1.0.7,<2.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "openai (>=2.16.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)"
]

[tool.poetry]
packages = [{include = "support_copilot", from = "."}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.0,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
